+++ device_calendar/android/src/main/kotlin/com/builttoroam/devicecalendar/models/Attendee.kt
@@ -1,4 +1,4 @@
 package com.builttoroam.devicecalendar.models
 
-class Attendee(val eventId: Long, val emailAddress: String, val name: String?, val isRequired: Boolean, val attendanceStatus: Int, val isOrganizer: Boolean) {
+class Attendee(var eventId: Long?, val emailAddress: String, val name: String?, val isRequired: Boolean?, val attendanceStatus: Int?, val isOrganizer: Boolean?) {
 }