+++ device_calendar/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Made event title optional to fix issue [72](https://github.com/builttoroam/flutter_plugins/issues/72)
 * Return information about the organiser of the event as per issue [73](https://github.com/builttoroam/flutter_plugins/issues/73)
 * Return attendance status of attendees and if they're required for an event. These are details are different across iOS and Android and so are returned within platform-specific objects
+* Ability to modify attendees for an event
 * **BREAKING CHANGE** `retrieveCalendars` and `retrieveEvents` now return lists that cannot be modified as part of address issue [113](https://github.com/builttoroam/flutter_plugins/issues/113)
 
 # 0.2.2 19th August 2019